move verbs to modular components
The new modular component format allows operators to implement their own
components. While the base components are powerful, user defined
components will allow directord to meet the needs individual operators.

An example user defined component has been added to the base "components"
directory.

The setup.py has been modified to ensure community defined components
added to base directord are installed into a share localtion.
