+++ .github/workflows/run-unit-test.yml
@@ -24,4 +24,4 @@ jobs:
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --show-source --max-complexity=23 --max-line-length=79 --statistics --benchmark --ignore=H104,W503
+        flake8 . --count --show-source --max-complexity=25 --max-line-length=79 --statistics --benchmark --ignore=H104,W503
