+++ directord/components/builtin_query.py
@@ -1,0 +1,70 @@
+#   Copyright Peznauts <kevin@cloudnull.com>. All Rights Reserved.
+#
+#   Licensed under the Apache License, Version 2.0 (the "License"); you may
+#   not use this file except in compliance with the License. You may obtain
+#   a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#   WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#   License for the specific language governing permissions and limitations
+#   under the License.
+
+import json
+
+from directord import components
+
+
+class Component(components.ComponentBase):
+    def __init__(self):
+        super().__init__(desc="Process cachefile commands")
+
+    def args(self):
+        super().args()
+        self.parser.add_argument(
+            "query",
+            help=(
+                "Scan the environment for a given cached argument and"
+                " store the resultant on the target. The resultant is"
+                " set in dictionary format: `{'client-id': ...}`"
+            ),
+        )
+
+    def server(self, exec_string, data, arg_vars):
+        """Return data from formatted transfer action.
+
+        :param exec_string: Inpute string from action
+        :type exec_string: String
+        :param data: Formatted data hash
+        :type data: Dictionary
+        :param arg_vars: Pre-Formatted arguments
+        :type arg_vars: Dictionary
+        :returns: Dictionary
+        """
+
+        super().server(exec_string=exec_string, data=data, arg_vars=arg_vars)
+        data["query"] = self.known_args.query
+        return data
+
+    def client(self, conn, cache, job):
+        """Run query command operation.
+
+        :param conn: Connection object used to store information used in a
+                     return message.
+        :type conn: Object
+        :param cache: Caching object used to template items within a command.
+        :type cache: Object
+        :param job: Information containing the original job specification.
+        :type job: Dictionary
+        :returns: tuple
+        """
+
+        super().client(conn=conn, cache=cache, job=job)
+        args = cache.get("args")
+        if args:
+            query = json.dumps(args.get(job["query"]))
+        else:
+            query = None
+        return query, None, True
