+++ orchestrations/files/test-blueprint.yaml.j2
@@ -1,6 +1,7 @@
 ---
 # Render runtime args
 test_stdout_arg: "{{ test_stdout_arg }}"
+test_stdout_arg2: "{{ test_stdout_arg2 }}"
 test_runtime_arg: "{{ test_runtime_arg }}"
 test_runtime_arg_as_env: "{{ test_runtime_arg_as_env }}"
